requirements.txt
fastapi
uvicorn
pillow
requests
httpx
python-multipart
celery
redis